chore: rename createGHA file
